backend/requirements.txt
fastapi
sqlmodel
bcrypt
passlib[bcrypt]
python-jose[cryptography]
python-multipart
uvicorn
psycopg2-binary
python-dotenv
email-validator
